Make `bazel run <hugo_website>` exec `hugo serve`
Unfortunately, this does not support live updates. By this I mean that
when changes are made to the Bazel source tree, these changes are not
automatically copied to the Hugo source tree, so Hugo doesn't rebuild
the website. Essentially, `bazel run` only serves the state of the
website when `bazel run` is executed.