Fix `bazel run` for `hugo_website` in root pkg
Previously, the `hugo serve` script was failing when a `hugo_website`
was declared in the root package (i.e. the unnamed package) because in
the generated script the path to the Hugo source directory included a
leading `/`. It was thus interpreted as an absolute path, though we want
a relative path.